Keep test_build_block inline for diff parity with compressed builder
Move only the extra unit tests to build_interned_block/unit_tests.rs;
the expensive integration test stays in the source file matching
build_compressed_block.rs structure.
